% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czestotliwosc.R
\name{czestotliwosc}
\alias{czestotliwosc}
\title{Funkcja zapisuje plik z napisami dla danego filmu,
nastepnie bada czestotliwosc wystepowania literek w tychze napisach}
\usage{
czestotliwosc(tytul, jezyk)
}
\arguments{
\item{tytul}{wektor jednoelementowy typu character - tytul filmu (jednowyrazowy)}

\item{jezyk}{wektor jednoelementowy typu character - jezyk filmu}
}
\description{
Funkcja \code{czestotliwosc(tytul, jezyk)} pobiera napisy ze strony http://www.opensubtitles.org/,
nastepnie bada czestotliwosc wystepowania literek w tychze napisach
}
\examples{
czestotliwosc("shrek", "eng")
}
\author{
Emilia Momotko, Martyna śpiewak
}

